feat: car speed according to the traffics and doing 900 x 900 grid
The speed is determined by the pervious speed +/- a random number of
normal distribution with zero mean and five sd. If a block is over 50%
full, the speed will be reduced. Else the speed will be increased. The
minimum speed is then bounded by zero. The maximum is bounded by road
type. Regular road is 30, big road is 50, and free way is 70. The grid
is also changed to the finest setting to check for performance issue.