fix: change some simulation logic to make it more understandable
The car exit check and the car capacity check is separated and each
road being checked for current number of exit cars and then whether new
location has the capacity. The simulator also marge the if statements
for different upper bound of the speed of roads. Checking for the
distance to end also changed to > 0 instead.